chore: Change Diractory Structure
